add delete validate id

+++ src/items/items.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, NotFoundException } from '@nestjs/common';
 import { CreateItemDto } from './dto/create-item.dto';
 import { UpdateItemDto } from './dto/update-item.dto';
 import { Repository } from 'typeorm';
@@ -39,8 +39,13 @@ export class ItemsService {
     return this.itemRepository.save({ id, ...updateItemDto });
   }
 
-  remove(id: number) {
+  async remove(id: number) {
     // const where = { id: id}
+    // find by id
+    const item = await this.itemRepository.findOneBy({ id })
+    if (!item) {
+      throw new NotFoundException(`Not found: id=${id}`)
+    }
     return this.itemRepository.delete({ id })
   }
 }
